refactor(*): move payload handling to its own module
